[Decomp][Spell] Cohesion split: extract effect delivery to SpellHit.cpp
